requirements.txt
anthropic==0.49.0
python-dotenv==1.0.1
manim==0.19.0
gtts==2.5.1
moviepy==1.0.3
